+++ lib/panda_doc/objects/base.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module PandaDoc
+  module Objects
+    class Base < Dry::Struct
+      transform_keys(&:to_sym)
+    end
+  end
+end