+++ lib/panda_doc/objects/session.rb
@@ -1,10 +1,10 @@
+# frozen_string_literal: true
+
 module PandaDoc
   module Objects
-    class Session
-      include Virtus.model
-
-      attribute :id,         String
-      attribute :expires_at, DateTime
+    class Session < Base
+      attribute :id, Types::Coercible::String
+      attribute :expires_at, Types::Params::DateTime
     end
   end
 end