+++ lib/panda_doc/response_factory.rb
@@ -1,28 +1,20 @@
+# frozen_string_literal: true
+
 module PandaDoc
   class ResponseFactory
     attr_reader :type
     private :type
+
+    def self.build(type)
+      new(type).build
+    end
 
     def initialize(type)
       @type = type.capitalize
     end
 
     def build
-      response_class.new(object_class.new)
-    end
-
-    private
-
-    def response_class
-      class_for("Responses")
-    end
-
-    def object_class
-      class_for("Objects")
-    end
-
-    def class_for(namespace)
-      Object.const_get("PandaDoc::#{namespace}::#{type}")
+      PandaDoc::Objects.const_get(type)
     end
   end
 end