+++ lib/panda_doc/types.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module PandaDoc
+  module Types
+    include Dry.Types()
+
+    module Custom
+      DocumentStatus = Dry::Types["nominal.string"].constructor(
+        Coercions.method(:to_splitted_string)
+      )
+    end
+  end
+end