+++ spec/document_spec.rb
@@ -1,9 +1,13 @@
+# frozen_string_literal: true
+
 require "helper"
 
 RSpec.shared_examples "a document object interface" do
   %w(
+    id
     uuid
     name
+    version
   ).each do |method|
     it "has #{method}" do
       expect(subject.public_send(method)).to eq(body[method])
@@ -26,6 +30,12 @@ RSpec.shared_examples "a document object interface" do
     ).to eq(to_seconds(body["date_modified"]))
   end
 
+  it "has expires_at" do
+    expect(
+      to_seconds(subject.expires_at)
+    ).to eq(to_seconds(body["expiration_date"]))
+  end
+
   %w(
     email
     first_name
@@ -39,6 +49,17 @@ RSpec.shared_examples "a document object interface" do
       ).to eq(body["recipients"].first[method])
     end
   end
+
+  %w(
+    name
+    value
+  ).each do |method|
+    it "has tokens.#{method}" do
+      expect(
+        subject.tokens.first.public_send(method)
+      ).to eq(body["tokens"].first[method])
+    end
+  end
 end
 
 RSpec.shared_examples "a failure result" do
@@ -50,6 +71,7 @@ RSpec.describe PandaDoc::Document do
   let(:successful_response) { double(success?: true, body: body) }
   let(:document_body) do
     {
+      "id" => "DOCUMENT_UUID",
       "uuid" => "DOCUMENT_UUID",
       "status" => "document.uploaded",
       "name" => "DOCUMENT_NAME",
@@ -63,7 +85,12 @@ RSpec.describe PandaDoc::Document do
         }
       ],
       "date_created" => "2014-10-06T08:42:13.836022Z",
-      "date_modified" => "2014-10-06T08:42:13.836048Z"
+      "date_modified" => "2014-10-06T08:42:13.836048Z",
+      "expiration_date" => "2021-02-22T00:51:59.474648Z",
+      "version" => "1",
+      "tokens" => [
+        { "name" => "token.name", "value" => "token value" }
+      ]
     }
   end
 
@@ -172,6 +199,29 @@ RSpec.describe PandaDoc::Document do
     before do
       allow(PandaDoc::ApiClient).to receive(:request)
         .with(:get, "/documents/uuid")
+        .and_return(response)
+    end
+
+    context "with failed response" do
+      let(:response) { failed_response }
+
+      it_behaves_like "a failure result"
+    end
+
+    context "with successful response" do
+      let(:response) { successful_response }
+      let(:body) { document_body }
+
+      it_behaves_like "a document object interface"
+    end
+  end
+
+  describe ".details" do
+    subject { described_class.details("uuid") }
+
+    before do
+      allow(PandaDoc::ApiClient).to receive(:request)
+        .with(:get, "/documents/uuid/details")
         .and_return(response)
     end